rag-perplexity-hackathon/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
chromadb>=0.5.0
sentence-transformers>=3.0.0
# Torch is typically required by sentence-transformers
torch>=2.2.0
requests>=2.32.0
pydantic>=2.7.0
python-dotenv>=1.0.1
python-multipart>=0.0.9
pdfplumber>=0.11.0
python-docx>=1.1.2
